+++ R/rfm-table-transaction.R
@@ -7,22 +7,19 @@
 #' @param order_date Date of the transaction.
 #' @param revenue Revenue from the customer.
 #' @param analysis_date Date of analysis.
-#' @param recency_bins Number of bins for recency.
-#' @param frequency_bins Number of bins for frequency.
-#' @param monetary_bins Number of bins for monetary.
+#' @param recency_bins Number of bins for recency or custom threshold.
+#' @param frequency_bins Number of bins for frequency or custom threshold.
+#' @param monetary_bins Number of bins for monetary or custom threshold.
 #' @param ... Other arguments.
 #'
-#' @return \code{rfm_table_order} returns a tibble with the following columns:
+#' @return \code{rfm_table_order} returns a list with the following:
 #'
-#' \item{customer_id}{Unique id of the customer.}
-#' \item{date_most_recent}{Date of the most recent transaction.}
-#' \item{recency_days}{Number of days since the most recent transaction.}
-#' \item{transaction_count}{Total number of transactions of the customer.}
-#' \item{amount}{Revenue from the customer.}
-#' \item{recency_score}{Recency score of the customer.}
-#' \item{frequency_score}{Frequency score of the customer.}
-#' \item{monetary_score}{Monetary score of the customer.}
-#' \item{rfm_score}{RFM score of the customer.}
+#' \item{rfm}{RFM table.}
+#' \item{analysis_date}{Date of analysis.}
+#' \item{frequency_bins}{Number of bins used for frequency score.}
+#' \item{recency_bins}{Number of bins used for recency score.}
+#' \item{monetary_bins}{Number of bins used for monetary score.}
+#' \item{threshold}{tibble with thresholds used for generating RFM scores.}
 #'
 #' @examples
 #' analysis_date <- lubridate::as_date('2006-12-31', tz = 'UTC')
@@ -31,6 +28,11 @@
 #' # access rfm table
 #' result <- rfm_table_order(rfm_data_orders, customer_id, order_date, revenue, analysis_date)
 #' result$rfm
+#'
+#' # using custom threshold
+#' rfm_table_order(rfm_data_orders, customer_id, order_date, revenue, analysis_date,
+#' recency_bins = c(115, 181, 297, 482), frequency_bins = c(4, 5, 6, 8),
+#' monetary_bins = c(256, 382, 506, 666))
 #'
 #' @export
 #'
@@ -69,10 +71,11 @@ rfm_table_order.default <- function(data = NULL, customer_id = NULL, order_date 
   result$frequency_score <- NA
   result$monetary_score  <- NA
 
-  rscore <-
-    recency_bins %>%
-    seq_len(.) %>%
-    rev()
+  if (length(recency_bins) == 1) {
+    rscore <- rev(seq_len(recency_bins))
+  } else {
+    rscore <- rev(seq_len((length(recency_bins) + 1)))
+  }
 
   if (length(recency_bins) == 1) {
     bins_recency <- bins(result, recency_days, recency_bins)
@@ -82,15 +85,19 @@ rfm_table_order.default <- function(data = NULL, customer_id = NULL, order_date 
   lower_recency <- bins_lower(result, recency_days, bins_recency)
   upper_recency <- bins_upper(result, recency_days, bins_recency)
 
-  for (i in seq_len(recency_bins)) {
+  rscore_len <- length(rscore)
+
+  for (i in seq_len(rscore_len)) {
     result$recency_score[result$recency_days >= lower_recency[i] &
       result$recency_days < upper_recency[i]] <- rscore[i]
   }
 
-  fscore <-
-    frequency_bins %>%
-    seq_len(.) %>%
-    rev()
+
+  if (length(frequency_bins) == 1) {
+    fscore <- rev(seq_len(frequency_bins))
+  } else {
+    fscore <- rev(seq_len((length(frequency_bins) + 1)))
+  }
 
   if (length(frequency_bins) == 1) {
     bins_frequency <- bins(result, transaction_count, frequency_bins)
@@ -100,15 +107,18 @@ rfm_table_order.default <- function(data = NULL, customer_id = NULL, order_date 
   lower_frequency <- bins_lower(result, transaction_count, bins_frequency)
   upper_frequency <- bins_upper(result, transaction_count, bins_frequency)
 
-  for (i in seq_len(frequency_bins)) {
+  fscore_len <- length(fscore)
+
+  for (i in seq_len(fscore_len)) {
     result$frequency_score[result$transaction_count >= lower_frequency[i] &
       result$transaction_count < upper_frequency[i]] <- i
   }
 
-  mscore <-
-    monetary_bins %>%
-    seq_len(.) %>%
-    rev()
+  if (length(monetary_bins) == 1) {
+    mscore <- rev(seq_len(monetary_bins))
+  } else {
+    mscore <- rev(seq_len((length(monetary_bins) + 1)))
+  }
 
   if (length(monetary_bins) == 1) {
     bins_monetary <- bins(result, amount, monetary_bins)
@@ -118,7 +128,9 @@ rfm_table_order.default <- function(data = NULL, customer_id = NULL, order_date 
   lower_monetary <- bins_lower(result, amount, bins_monetary)
   upper_monetary <- bins_upper(result, amount, bins_monetary)
 
-  for (i in seq_len(monetary_bins)) {
+  mscore_len <- length(mscore)
+
+  for (i in seq_len(mscore_len)) {
     result$monetary_score[result$amount >= lower_monetary[i] &
       result$amount < upper_monetary[i]] <- i
   }
@@ -134,12 +146,20 @@ rfm_table_order.default <- function(data = NULL, customer_id = NULL, order_date 
 
   result$transaction_count <- as.numeric(result$transaction_count)
 
+  threshold <- tibble::tibble(recency_lower   = lower_recency,
+                              recency_upper   = upper_recency,
+                              frequency_lower = lower_frequency,
+                              frequency_upper = upper_frequency,
+                              monetary_lower  = lower_monetary,
+                              monetary_upper  = upper_monetary)
+
   out <- list(
     rfm            = result,
     analysis_date  = analysis_date,
     frequency_bins = frequency_bins,
     recency_bins   = recency_bins,
-    monetary_bins  = monetary_bins
+    monetary_bins  = monetary_bins,
+    threshold      = threshold
   )
 
   class(out) <- c("rfm_table_order", "tibble", "data.frame")